extract as a module
